checklist updated portion

+++ database/migrations/2019_06_12_112354_add_category_column_to_checklists_table.php
@@ -25,8 +25,12 @@ class AddCategoryColumnToChecklistsTable extends Migration
      */
     public function down()
     {
-        Schema::table('checklists', function (Blueprint $table) {
-            $table->dropColumn('category');
-        });
+        if (Schema::hasColumn('checklists', 'category'))
+        {            
+            Schema::table('checklists', function (Blueprint $table)
+            {
+                $table->dropColumn('category');
+            });
+        }
     }
 }
